Revert "libcamera: ipu3: imgu: Add pipe calculation debug"
This reverts commit 5b015e96ccdbcd87b4ba6484199652fec5cdb38a.

The ImgU pipe configuration debug is useful to test the correctness
of the parameters computation against the Intel Python script.

However, the number of debug messages which is printed out by the
configuration procedure is so high it floods the logs, up to the point
that starting the Android camera3 HAL, which tests several configurations
at startup, becomes so slow it is barely usable.

Revert the patch that adds the excessive debug statements, which are mostly
useful only when testing the configuration procedure.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Acked-by: Umang Jain <umang.jain@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>